created usePlotDataHandling hook |  updated useHoverHandlers name to usePlotInteractionHandlers
